move step-vanilla to common/app
